requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
google-genai>=1.0.0
pgvector>=0.2.5
aiohttp>=3.9.0
cachetools>=5.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
google-cloud-storage>=2.16.0
google-cloud-secret-manager>=2.20.0
tenacity>=8.2.0
structlog>=24.1.0